feat(fix): correct pr
