+++ src/modules/client/create-announcement/hooks/use-additional-info.tsx
@@ -1,0 +1,36 @@
+import type { TxKeyPath } from '@/core';
+import { useCustomForm } from '@/core';
+import { useFormStepper } from '@/shared';
+import type { AnnouncementType } from '@/types/announcement';
+
+import { CreateAnnouncementStepTenSchema } from '../schemas';
+import type { projectExtensionsFormType } from '../types';
+
+export const useAdditionalInfo = () => {
+  const { onHandleBack, onHandleNext, setFormData, formData } =
+    useFormStepper<AnnouncementType>();
+  const { handleSubmit, control, errors } = useCustomForm(
+    CreateAnnouncementStepTenSchema,
+    { projectExtensions: formData?.projectExtensions || [] }
+  );
+  const error = errors?.projectExtensions?.message as TxKeyPath | undefined;
+
+  const onSubmit = (data: projectExtensionsFormType) => {
+    setFormData((prev: any) => ({
+      ...prev,
+      ...data,
+    }));
+    onHandleNext();
+  };
+
+  return {
+    onHandleBack,
+    onHandleNext,
+    setFormData,
+    formData,
+    handleSubmit,
+    control,
+    error,
+    onSubmit,
+  };
+};
