+++ src/modules/client/create-announcement/hooks/use-property-type.tsx
@@ -1,0 +1,37 @@
+import type { TxKeyPath } from '@/core';
+import { useCustomForm } from '@/core';
+import { useFormStepper } from '@/shared';
+import type { AnnouncementType } from '@/types/announcement';
+
+import { CreateAnnouncementStepFourSchema } from '../schemas';
+import type { propertyTypeFormType } from '../types';
+
+export const usePropertyType = () => {
+  const { onHandleBack, onHandleNext, setFormData, formData } =
+    useFormStepper<AnnouncementType>();
+
+  const { handleSubmit, control, errors } = useCustomForm(
+    CreateAnnouncementStepFourSchema,
+    { propertyType: formData?.propertyType }
+  );
+  const error = errors?.propertyType?.message as TxKeyPath | undefined;
+
+  const onSubmit = (data: propertyTypeFormType) => {
+    setFormData((prev: any) => ({
+      ...prev,
+      ...data,
+    }));
+    onHandleNext();
+  };
+
+  return {
+    onHandleBack,
+    onHandleNext,
+    setFormData,
+    formData,
+    handleSubmit,
+    control,
+    error,
+    onSubmit,
+  };
+};
