+++ src/modules/client/create-announcement/hooks/use-speciality.tsx
@@ -1,0 +1,35 @@
+import type { TxKeyPath } from '@/core';
+import { useCustomForm } from '@/core';
+import { useFormStepper } from '@/shared';
+import type { AnnouncementType } from '@/types/announcement';
+
+import { CreateAnnouncementStepOneSchema } from '../schemas';
+import type { SpecialityFormType } from '../types';
+
+export const useSpeciality = () => {
+  const { onHandleBack, onHandleNext, setFormData, formData } =
+    useFormStepper<AnnouncementType>();
+  const { handleSubmit, control, errors } = useCustomForm(
+    CreateAnnouncementStepOneSchema,
+    { architectSpeciality: formData.architectSpeciality }
+  );
+  const error = errors?.architectSpeciality?.message as TxKeyPath | undefined;
+
+  const onSubmit = (data: SpecialityFormType) => {
+    setFormData((prev: any) => ({
+      ...prev,
+      ...data,
+    }));
+    onHandleNext();
+  };
+  return {
+    onHandleBack,
+    onHandleNext,
+    setFormData,
+    formData,
+    handleSubmit,
+    control,
+    error,
+    onSubmit,
+  };
+};
