+++ src/modules/client/create-announcement/types/index.ts
@@ -1,5 +1,7 @@
 import type { SvgProps } from 'react-native-svg';
 
+import type { AnnouncementType } from '@/types/announcement';
+
 export type sharedType = {
   id: number;
   label: string;
@@ -14,3 +16,27 @@ export type workType = {
   header: string;
   description: string;
 };
+
+export type SpecialityFormType = Pick<AnnouncementType, 'architectSpeciality'>;
+export type needsFormType = Pick<AnnouncementType, 'needs'>;
+export type workTypeFormType = Pick<AnnouncementType, 'workType'>;
+export type architecturalStyleFormType = Pick<
+  AnnouncementType,
+  'architecturalStyle'
+>;
+export type projectCategoryFormType = Pick<AnnouncementType, 'projectCategory'>;
+export type projectExtensionsFormType = Pick<
+  AnnouncementType,
+  'projectExtensions'
+>;
+export type propertyTypeFormType = Pick<AnnouncementType, 'propertyType'>;
+export type piecesRenovateFormType = Pick<AnnouncementType, 'piecesRenovate'>;
+export type CreateProfileFormType = Pick<
+  AnnouncementType,
+  | 'firstName'
+  | 'lastName'
+  | 'email'
+  | 'phoneNumber'
+  | 'rules'
+  | 'receiveNotifications'
+>;
