+++ src/translations/ar/validations.json
@@ -1,19 +1,19 @@
 {
   "required": "مطلوب",
   "invalid": "تنسيق البريد الإلكتروني غير صالح",
-  "email-max-length": "يجب أن يحتوي البريد الإلكتروني على 100 حرف كحد أقصى",
-  "password-min-length": "يجب أن تحتوي كلمة المرور على 8 أحرف على الأقل",
-  "password-lowercase": "يجب أن تحتوي كلمة المرور على حرف صغير واحد على الأقل",
-  "password-uppercase": "يجب أن تحتوي كلمة المرور على حرف كبير واحد على الأقل",
-  "password-digit": "يجب أن تحتوي كلمة المرور على رقم واحد على الأقل",
-  "password-special-char": "يجب أن تحتوي كلمة المرور على حرف خاص واحد على الأقل",
-  "field-min-length": "يجب أن يحتوي الحقل على 3 أحرف كحد أدنى",
-  "field-max-length": "يجب أن يحتوي الحقل على 50 حرف كحد أقصى",
-  "field-shape": "يجب أن يتبع الحقل التنسيق المطلوب",
-  "confirm-password": "يجب أن تتطابق كلمات المرور",
-  "phone-number-digits": "يجب أن يحتوي رقم الهاتف على أرقام فقط",
-  "phone-number-min-length": "يجب أن يحتوي رقم الهاتف على 8 أحرف على الأقل",
-  "phone-number-max-length": "يجب أن يحتوي رقم الهاتف على 15 حرفًا كحد أقصى",
+  "emailMaxLength": "يجب أن يحتوي البريد الإلكتروني على 100 حرف كحد أقصى",
+  "passwordMinLength": "يجب أن تحتوي كلمة المرور على 8 أحرف على الأقل",
+  "passwordLowercase": "يجب أن تحتوي كلمة المرور على حرف صغير واحد على الأقل",
+  "passwordUppercase": "يجب أن تحتوي كلمة المرور على حرف كبير واحد على الأقل",
+  "passwordDigit": "يجب أن تحتوي كلمة المرور على رقم واحد على الأقل",
+  "passwordSpecialChar": "يجب أن تحتوي كلمة المرور على حرف خاص واحد على الأقل",
+  "fieldMinLength": "يجب أن يحتوي الحقل على 3 أحرف كحد أدنى",
+  "fieldMaxLength": "يجب أن يحتوي الحقل على 50 حرف كحد أقصى",
+  "fieldShape": "يجب أن يتبع الحقل التنسيق المطلوب",
+  "confirmPassword": "يجب أن تتطابق كلمات المرور",
+  "phoneNumberDigits": "يجب أن يحتوي رقم الهاتف على أرقام فقط",
+  "phoneNumberMinLength": "يجب أن يحتوي رقم الهاتف على 8 أحرف على الأقل",
+  "phoneNumberMaxLength": "يجب أن يحتوي رقم الهاتف على 15 حرفًا كحد أقصى",
   "arrayNonEmptyString": "يجب عليك اختيار خيار واحد على الأقل",
   "uploadedImageType": ".الرجاء تحميل ملفات الصور فقط",
   "arrayNonEmptyImage": "يجب عليك رفع صورة واحدة على الأقل"
