+++ src/translations/en/validations.json
@@ -1,19 +1,19 @@
 {
   "required": "Required",
   "invalid": "Invalid email format",
-  "email-max-length": "Email must be at most 100 characters",
-  "password-min-length": "Password must be at least 8 characters",
-  "password-lowercase": "Password must contain at least one lowercase letter",
-  "password-uppercase": "Password must contain at least one uppercase letter",
-  "password-digit": "Password must contain at least one digit",
-  "password-special-char": "Password must contain at least one special character",
-  "field-min-length": "Field must be at least 3 characters",
-  "field-max-length": "Field must be at most 50 characters",
-  "field-shape": "Field must follow the required format",
-  "confirm-password": "Passwords must match",
-  "phone-number-digits": "Phone number must contain only digits",
-  "phone-number-min-length": "Phone number must be at least 8 characters",
-  "phone-number-max-length": "Phone number must be at most 15 characters",
+  "emailMaxLength": "Email must be at most 100 characters",
+  "passwordMinLength": "Password must be at least 8 characters",
+  "passwordLowercase": "Password must contain at least one lowercase letter",
+  "passwordUppercase": "Password must contain at least one uppercase letter",
+  "passwordDigit": "Password must contain at least one digit",
+  "passwordSpecialChar": "Password must contain at least one special character",
+  "fieldMinLength": "Field must be at least 3 characters",
+  "fieldMaxLength": "Field must be at most 50 characters",
+  "fieldShape": "Field must follow the required format",
+  "confirmPassword": "Passwords must match",
+  "phoneNumberDigits": "Phone number must contain only digits",
+  "phoneNumberMinLength": "Phone number must be at least 8 characters",
+  "phoneNumberMaxLength": "Phone number must be at most 15 characters",
   "arrayNonEmptyNumbers": "You have to pick at least one option",
   "uploadedImageType": "Please upload only image files.",
   "arrayNonEmptyImage": "You have to upload at least one image"
