+++ src/translations/fr/validations.json
@@ -1,19 +1,19 @@
 {
   "required": "Requis",
   "invalid": "Format d'email invalide",
-  "email-max-length": "Le courriel doit contenir au plus 100 caractères",
-  "password-min-length": "Le mot de passe doit contenir au moins 8 caractères",
-  "password-lowercase": "Le mot de passe doit contenir au moins une lettre minuscule",
-  "password-uppercase": "Le mot de passe doit contenir au moins une lettre majuscule",
-  "password-digit": "Le mot de passe doit contenir au moins un chiffre",
-  "password-special-char": "Le mot de passe doit contenir au moins un caractère spécial",
-  "field-min-length": "Le champ doit contenir au moins 3 caractères",
-  "field-max-length": "Le champ doit contenir au plus 50 caractères",
-  "field-shape": "Le champ doit respecter le format requis",
-  "confirm-password": "Les mots de passe doivent correspondre",
-  "phone-number-digits": "Le numéro de téléphone doit contenir uniquement des chiffres",
-  "phone-number-min-length": "Le numéro de téléphone doit comporter au moins 8 caractères",
-  "phone-number-max-length": "Le numéro de téléphone doit comporter au maximum 15 caractères",
+  "emailMaxLength": "Le courriel doit contenir au plus 100 caractères",
+  "passwordMinLength": "Le mot de passe doit contenir au moins 8 caractères",
+  "passwordLowercase": "Le mot de passe doit contenir au moins une lettre minuscule",
+  "passwordUppercase": "Le mot de passe doit contenir au moins une lettre majuscule",
+  "passwordDigit": "Le mot de passe doit contenir au moins un chiffre",
+  "passwordSpecialChar": "Le mot de passe doit contenir au moins un caractère spécial",
+  "fieldMinLength": "Le champ doit contenir au moins 3 caractères",
+  "fieldMaxLength": "Le champ doit contenir au plus 50 caractères",
+  "fieldShape": "Le champ doit respecter le format requis",
+  "confirmPassword": "Les mots de passe doivent correspondre",
+  "phoneNumberDigits": "Le numéro de téléphone doit contenir uniquement des chiffres",
+  "phoneNumberMinLength": "Le numéro de téléphone doit comporter au moins 8 caractères",
+  "phoneNumberMaxLength": "Le numéro de téléphone doit comporter au maximum 15 caractères",
   "arrayNonEmptyString": "Vous devez choisir au moins une option",
   "uploadedImageType": "Veuillez télécharger uniquement des fichiers image.",
   "arrayNonEmptyImage": "Vous devez charger au moins une image"
